added line to workaround makefile issue on John's laptop; changed to floats from doubles; implemented work with Adam's config stuff

+++ include/statistics.h
@@ -8,9 +8,9 @@ extern std::mt19937 GENERATOR;
 
 // Distribution template used to sample from random distributions
 template <typename Distribution>
-std::vector<double> sampleFromDistribution(size_t n, Distribution& distribution) {
+std::vector<float> sampleFromDistribution(size_t n, Distribution& distribution) {
     // Generate samples
-    std::vector<double> samples(n);
+    std::vector<float> samples(n);
     for (size_t i = 0; i < n; ++i) {
         samples[i] = distribution(GENERATOR);
     }
